Move `BloopConfig` to shared core
Make sure that we have access to this from different integrations.
