+++ integrations/sbt-bloop/src/main/scala-sbt-1.0/bloop/DiscoveredSbtPlugins.scala
@@ -1,4 +1,4 @@
-package bloop
+package bloop.integrations.sbt
 
 import sbt.Setting
