+++ integrations/sbt-bloop/src/main/scala/bloop/SbtBloop.scala
@@ -1,9 +1,7 @@
-package bloop
+package bloop.integrations.sbt
 
-import java.io.FileOutputStream
-import java.io.File
-
-import sbt.{Keys, AutoPlugin, Def, ScopeFilter, ThisBuild, Compile, Test, Configuration}
+import bloop.integrations.BloopConfig
+import sbt.{Keys, AutoPlugin, Def, ScopeFilter, ThisBuild, Compile, Test, Configuration, File}
 
 object SbtBloop extends AutoPlugin {
   import sbt.plugins.JvmPlugin
@@ -36,51 +34,6 @@ object PluginImplementation {
     inConfig(conf) {
       List(bloopGenerate := PluginDefaults.bloopGenerate.value) ++
         DiscoveredSbtPlugins.settings // discoveredSbtPlugins triggers compilation in 0.13, we replace it.
-    }
-  }
-
-  case class Config(
-      name: String,
-      baseDirectory: File,
-      dependencies: Seq[String],
-      scalaOrganization: String,
-      scalaName: String,
-      scalaVersion: String,
-      classpath: Seq[File],
-      classesDir: File,
-      scalacOptions: Seq[String],
-      javacOptions: Seq[String],
-      sourceDirectories: Seq[File],
-      testFrameworks: Seq[Seq[String]],
-      fork: Boolean,
-      javaHome: File,
-      javaOptions: Seq[String],
-      allScalaJars: Seq[File],
-      tmp: File
-  ) {
-    private def seqToString[T](xs: Seq[T], sep: String = ","): String = xs.mkString(sep)
-    private def toPaths(xs: Seq[File]): Seq[String] = xs.map(_.getAbsolutePath)
-    def toProperties: java.util.Properties = {
-      val properties = new java.util.Properties()
-      properties.setProperty("name", name)
-      properties.setProperty("baseDirectory", baseDirectory.getAbsolutePath)
-      properties.setProperty("dependencies", seqToString(dependencies))
-      properties.setProperty("scalaOrganization", scalaOrganization)
-      properties.setProperty("scalaName", scalaName)
-      properties.setProperty("scalaVersion", scalaVersion)
-      properties.setProperty("classpath", seqToString(toPaths(classpath)))
-      properties.setProperty("classesDir", classesDir.getAbsolutePath)
-      properties.setProperty("scalacOptions", seqToString(scalacOptions, ";"))
-      properties.setProperty("javacOptions", seqToString(javacOptions, ";"))
-      properties.setProperty("sourceDirectories", seqToString(toPaths(sourceDirectories)))
-      properties.setProperty("testFrameworks",
-                             seqToString(testFrameworks.map(seqToString(_)), sep = ";"))
-      properties.setProperty("fork", fork.toString)
-      properties.setProperty("javaHome", javaHome.getAbsolutePath)
-      properties.setProperty("javaOptions", seqToString(javaOptions, ";"))
-      properties.setProperty("allScalaJars", seqToString(toPaths(allScalaJars)))
-      properties.setProperty("tmp", tmp.getAbsolutePath)
-      properties
     }
   }
 
@@ -133,13 +86,11 @@ object PluginImplementation {
       val _ = Keys.copyResources.value
 
       // format: OFF
-      val config = Config(projectName, baseDirectory, dependenciesAndAggregates, scalaOrg,
-        scalaName,scalaVersion, classpath, classesDir, scalacOptions, javacOptions, sourceDirs,
+      val config = BloopConfig(projectName, baseDirectory, dependenciesAndAggregates, scalaOrg,
+        scalaName, scalaVersion, classpath, classesDir, scalacOptions, javacOptions, sourceDirs,
         testFrameworks, fork, javaHome, javaOptions, allScalaJars, tmp)
       sbt.IO.createDirectory(bloopConfigDir)
-      val stream = new FileOutputStream(outFile)
-      try config.toProperties.store(stream, null)
-      finally stream.close()
+      config.writeTo(outFile)
       logger.success(s"Bloop wrote the configuration of project '$projectName' to '$outFile'.")
       // format: ON
     }
